# openapi.yaml
openapi: 3.0.3
info:
  title: DRP · Domestic Resource Planning API
  description: >
    API REST del core de DRP (Domestic Resource Planning).


    Este documento es la **fuente de verdad** de la API, no una ilustración
    de ella (ver ADR-007): se edita antes que el código que lo implementa, el
    frontend genera de él sus tipos y su cliente, y el backend se verifica
    contra él con las pruebas de contrato. Ver readme.md del proyecto para el
    contexto completo — modelo de dominio, event bus y roadmap.


    Dos convenciones transversales. Los **errores de forma** responden 400 con
    código `VALIDATION_ERROR`; los de **regla de negocio**, 409 con el código
    concreto. Y todas las colecciones se devuelven **paginadas**, con la misma
    envoltura `{ items, page, size, total }`.
  version: "0.3.0-fase2"
  license:
    name: Ver LICENSE del repositorio
servers:
  - url: http://localhost:8080/api/v1
    description: Desarrollo local, levantado por compose.yaml
  - url: https://api.drp.example/api/v1
    description: Entorno ilustrativo (el real se fija al elegir el VPS, al cerrar la Fase 1)
...
security:
  - bearerAuth: []
paths:
  /households:
    post:
      operationId: createHousehold
      tags: [households]
      summary: Crear un hogar con su primer administrador
      description: >
        Unica escritura sin autenticar de la API. Crea la identidad sin
        verificar, el hogar, la pertenencia HOUSEHOLD_ADMIN y las categorias
        por defecto en una transaccion, y envia por correo un token de
        verificacion de un solo uso. NO devuelve sesion -- el hogar no es
        utilizable hasta verificar. Responde siempre 202, exista o no ya ese
        correo: contestar otra cosa permitiria averiguar quien esta
        registrado.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HouseholdInput" }
      responses:
        "202":
          description: Solicitud aceptada. Sin cuerpo, y siempre la misma respuesta.
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        default: { $ref: "#/components/responses/ServerError" }
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verificar el correo y obtener sesion
      description: >
        Consume el token de un solo uso enviado por correo, marca la identidad
        como verificada y devuelve el par de tokens. Si era el alta de un
        hogar, es aqui donde se publica HouseholdCreated -- no al insertar la
        fila, para que ningun modulo siembre datos de un hogar que quiza no
        llegue a usarse.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: Correo verificado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "409":
          description: Token caducado, ya usado o inexistente (VERIFICATION_TOKEN_INVALID)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        default: { $ref: "#/components/responses/ServerError" }
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Reenviar el enlace de verificacion
      description: >
        Invalida el token anterior y emite uno nuevo. Responde siempre igual,
        exista o no la identidad, por el mismo motivo que el alta de hogar.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202":
          description: Solicitud aceptada. Sin cuerpo, y siempre la misma respuesta.
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        default: { $ref: "#/components/responses/ServerError" }
  /auth/password-reset:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Solicitar el restablecimiento de contrasena
      description: >
        Sin credencial de ningun tipo. Si hay una identidad activa detras del
        correo, invalida su token de restablecimiento anterior, emite uno
        nuevo con una hora de caducidad y lo envia. Responde siempre 202,
        exista o no ese correo y este o no dada de baja la identidad:
        contestar otra cosa permitiria averiguar quien usa el sistema.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202":
          description: Solicitud aceptada. Sin cuerpo, y siempre la misma respuesta.
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        default: { $ref: "#/components/responses/ServerError" }
  /auth/password-reset/confirm:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Fijar la contrasena nueva con el token recibido
      description: >
        Consume el token de un solo uso, fija la contrasena y marca el correo
        como verificado si no lo estaba -- recibir el token prueba el control
        de la direccion. Revoca TODOS los refresh tokens de la identidad
        ANTES de emitir el par nuevo: si el motivo del restablecimiento era un
        acceso ajeno, dejar sesiones abiertas anularia el gesto.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword:
                  type: string
                  format: password
                  minLength: 12
      responses:
        "200":
          description: Contrasena restablecida; se devuelve sesion nueva
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "409":
          description: Token caducado, ya usado o inexistente (RESET_TOKEN_INVALID)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        default: { $ref: "#/components/responses/ServerError" }
  /auth/password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Cambiar la contrasena estando autenticado
      description: >
        Exige la contrasena actual ademas de la nueva: sin ese requisito, un
        access token robado bastaria para cambiarla y dejar fuera al dueno de
        la cuenta. Revoca las DEMAS sesiones y conserva la que esta en uso.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword:
                  type: string
                  format: password
                  minLength: 12
      responses:
        "204":
          description: Contrasena cambiada
        "401":
          description: La contrasena actual no es correcta (CURRENT_PASSWORD_INVALID)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        default: { $ref: "#/components/responses/ServerError" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Iniciar sesión (usuarios del hogar)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password:
                  type: string
                  format: password
                  description: >
                    Sin minimo declarado a proposito: aqui se comprueba una
                    credencial existente, no se fija una nueva. Rechazarla por
                    forma daria 400 donde debe darse 401."
      responses:
        "200":
          description: Autenticación correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: >
            Credenciales inválidas, o correo sin verificar
            (EMAIL_NOT_VERIFIED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        default: { $ref: "#/components/responses/ServerError" }
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [auth]
      summary: Renovar el access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: Nuevo par de tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Refresh token inválido o revocado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        default: { $ref: "#/components/responses/ServerError" }
  /auth/logout:
    post:
      operationId: revokeSession
      tags: [auth]
      summary: Cerrar la sesión en curso
      description: >
        Revoca el refresh token indicado, que es lo único revocable: un access
        token no se puede retirar y sigue valiendo hasta que caduque, unos 15
        minutos. Para tumbar todas las sesiones a la vez está el cambio de
        contraseña. Responde igual con un token válido que con uno que ya no
        existe: no hay nada que averiguar por la respuesta.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "204":
          description: Sesión cerrada
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Listar miembros del hogar
      parameters:
        - name: includeDeactivated
          in: query
          schema: { type: boolean, default: false }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Listado de usuarios activos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /users/me/avatar:
    put:
      operationId: setIdentityAvatar
      tags: [users]
      summary: Subir o sustituir el avatar propio
      description: >
        "me" resuelve a la IDENTIDAD del token, no a la pertenencia: el avatar
        es de la persona y sobrevive a cualquier hogar. Por eso NO es un fichero
        del hogar -- no suma a su gigabyte ni tiene politica de RLS detras (ver
        readme.md 4.1.1). Sustituye siempre al anterior, asi que no acumula.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Solo imagen, maximo 1 MB
      responses:
        "204":
          description: Avatar guardado
        "413":
          description: El fichero pasa de 1 MB (FILE_TOO_LARGE)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "415":
          description: El tipo real no es una imagen admitida (FILE_TYPE_NOT_ALLOWED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        default: { $ref: "#/components/responses/ServerError" }
    delete:
      operationId: deleteIdentityAvatar
      tags: [users]
      summary: Quitar el avatar propio
      responses:
        "204":
          description: Avatar retirado
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /invitations:
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: Listar las invitaciones vivas (solo HOUSEHOLD_ADMIN)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Invitaciones ni aceptadas, ni revocadas, ni caducadas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
    post:
      operationId: inviteUser
      tags: [invitations]
      summary: Invitar a alguien al hogar (solo HOUSEHOLD_ADMIN)
      description: >
        Emite un token de un solo uso con 7 dias de caducidad y lo envia por
        correo. No crea identidad ni pertenencia: eso ocurre al aceptar.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationInput" }
      responses:
        "201":
          description: Invitacion emitida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invitation" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: >
            Ya es miembro activo del hogar (ALREADY_MEMBER) o ya hay una
            invitacion viva para ese correo (INVITATION_ALREADY_PENDING)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Aceptar una invitacion
      description: >
        Sin autenticar -- lo autoriza el token recibido por correo. Crea la
        identidad ya VERIFICADA, porque haber recibido el token prueba el
        control de esa direccion, o vincula la que ya exista. name y password
        solo hacen falta si la identidad no existia.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationAcceptInput" }
      responses:
        "200":
          description: Invitacion aceptada; se devuelve sesion
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "409":
          description: >
            Token caducado, ya usado o revocado (INVITATION_TOKEN_INVALID), o
            la identidad ya pertenece a otro hogar (IDENTITY_ALREADY_MEMBER)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        default: { $ref: "#/components/responses/ServerError" }
  /invitations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: revokeInvitation
      tags: [invitations]
      summary: Revocar una invitacion (solo HOUSEHOLD_ADMIN)
      responses:
        "204":
          description: Invitacion revocada
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: deactivateUser
      tags: [users]
      summary: Dar de baja a un usuario del hogar (solo HOUSEHOLD_ADMIN)
      description: >
        Baja lógica -- marca deactivatedAt, revoca los refresh tokens del
        usuario e impide que vuelva a autenticarse. Sus assets quedan SIN
        PROPIETARIO, no se reasignan: se localizan después con
        GET /assets?withoutOwner=true. Sus préstamos se conservan. Publica
        UserDeactivated.
      responses:
        "204":
          description: Usuario dado de baja
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Es el único HOUSEHOLD_ADMIN activo del hogar
            (USER_LAST_ADMIN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /users/{id}/roles:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: changeUserRole
      tags: [users]
      summary: Modificar el rol estructural de un usuario (solo HOUSEHOLD_ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: "#/components/schemas/UserRole" }
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Dejaría al hogar sin ningún HOUSEHOLD_ADMIN activo
            (USER_LAST_ADMIN). Un hogar sin administrador no puede invitar,
            cambiar roles ni dar de baja a nadie: quedaría bloqueado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT de usuarios del hogar. Claims -- sub (identityId), memberId
        (pertenencia en el hogar), householdId, role y sid. La distincion entre
        sub y memberId importa: todo lo que el dominio llama "usuario"
        --propietario de un asset, prestador, receptor, createdBy-- apunta a la
        pertenencia, mientras que los refresh tokens cuelgan de la identidad.


        sid identifica cual de las sesiones de esa persona es esta: es el
        identificador del refresh token con el que se emitio. Existe porque
        POST /auth/password revoca las DEMAS sesiones y conserva la que esta en
        uso, y para conservarla hay que saber cual es. La alternativa era pedir
        el refresh token en el cuerpo de esa operacion, que obligaria al cliente
        a mandar una credencial de vida larga donde no hace falta. No es un
        secreto: identifica una fila cuyo valor real esta hasheado.
    loanToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Credencial de un prestador o receptor **externo**, sin cuenta en DRP.
        Es un JWT firmado **sin sub de usuario**, con claims loanId y role
        (LENDER | BORROWER), que viaja como enlace por correo al iniciar el
        prestamo.


        Su alcance son exactamente dos operaciones -- GET /loans/{id} de ESE
        prestamo y POST /loans/{id}/return -- y ninguna otra cosa del hogar.
        Ni el listado, ni otro prestamo, ni los assets.


        Vale a la vez la firma y la fila: su hash se guarda en
        loan_access_tokens, y verificarlo exige firma valida **y** fila viva.
        Las dos capas hacen cosas distintas y por eso estan las dos -- la firma
        lo hace infalsificable y deja rechazar lo caducado sin tocar la base de
        datos; la fila lo hace **revocable**, que un JWT por si solo no puede
...
  parameters:
...
    Page:
      name: page
      in: query
      description: Pagina solicitada, empezando en 0
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      description: Elementos por pagina
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
...
    Forbidden:
      description: No autorizado para esta operación
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Falta el token, ha caducado o no es válido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: >
        El cuerpo o los parámetros no cumplen el contrato (VALIDATION_ERROR).
        Es el error de forma; los de negocio van en 409. `details` lleva un
        campo por atributo rechazado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: >
        Se ha superado el límite de frecuencia (RATE_LIMITED). Se aplica sobre
        los endpoints sin autenticar y sobre la subida de ficheros.
      headers:
        Retry-After:
          description: Segundos que conviene esperar antes de reintentar
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: >
        Error no previsto (INTERNAL_ERROR). El cuerpo nunca lleva detalle
        interno: la traza va al log, no a la respuesta.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    TokenPair:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, example: 900 }
...
    HouseholdInput:
      type: object
      required: [name, timeZone, admin]
      properties:
        name: { type: string }
        timeZone:
          type: string
          example: Europe/Madrid
          description: Identificador IANA. Lo usa el proceso diario de vencidos
        admin:
          type: object
          required: [name, email, password]
          properties:
            name: { type: string }
            email: { type: string, format: email }
            password:
              type: string
              format: password
              minLength: 12
...
    InvitationInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
          description: Normalizado a minusculas, como en la identidad
        role: { $ref: "#/components/schemas/UserRole" }
...
    InvitationPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Invitation" }
...
    UserPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/User" }